Given data and it's associated changepoint data, annotates data points with the changepoint interval in which that data falls

Given data and it's associated changepoint data, annotates data points with the changepoint interval in which that data falls

annotate_data_with_changepoint(data_input, changepoint_data)

Arguments

data_input
The data associated with the given changepoints

changepoint_data
The changepoints, in a dataframe format (see generate_changepoint)

Value

A dataframe of the data_input annotated with associated changepoint

Examples

data <- data.frame(X = 1:10, Y = gfpop::dataGenerator(10, c(0.1, 0.3, 0.5, 0.8, 1),
c(1, 2, 1, 3, 1), sigma = 1))
graph <- gfpop::graph(type = "std")
changepoint_data <- generate_changepoint(data$Y, graph)
annotate_data_with_changepoint(data, changepoint_data)

#>     X          Y changepoint changepoint_end          y
#> 1   1 -0.4000435           1               0 -0.4000435
#> 2   2  2.2553171           2               1  2.2553171
#> 3   3 -0.4372636           3               2 -0.4372636
#> 4   4  0.9944287           4               3  0.9944287
#> 5   5  1.6215527           5               4  1.6215527
#> 6   6  4.1484116           6               5  4.1484116
#> 7   7  1.1781823           7               6  1.1781823
#> 8   8  2.7526747           8               7  2.7526747
#> 9   9  0.7558004           9               8  0.7558004
#> 10 10  0.7172946          10               9  0.7172946
#>                                                                   CP_Data
#> 1  Assoc. CP:  \n \t Begin: 0 \n \t End: 1 \n \t CP_Y: -0.400043516721755
#> 2    Assoc. CP:  \n \t Begin: 1 \n \t End: 2 \n \t CP_Y: 2.25531705484526
#> 3  Assoc. CP:  \n \t Begin: 2 \n \t End: 3 \n \t CP_Y: -0.437263611219531
#> 4   Assoc. CP:  \n \t Begin: 3 \n \t End: 4 \n \t CP_Y: 0.994428713253839
#> 5    Assoc. CP:  \n \t Begin: 4 \n \t End: 5 \n \t CP_Y: 1.62155272141521
#> 6    Assoc. CP:  \n \t Begin: 5 \n \t End: 6 \n \t CP_Y: 4.14841160602606
#> 7    Assoc. CP:  \n \t Begin: 6 \n \t End: 7 \n \t CP_Y: 1.17818233902337
#> 8    Assoc. CP:  \n \t Begin: 7 \n \t End: 8 \n \t CP_Y: 2.75267469792648
#> 9   Assoc. CP:  \n \t Begin: 8 \n \t End: 9 \n \t CP_Y: 0.755800393221617
#> 10 Assoc. CP:  \n \t Begin: 9 \n \t End: 10 \n \t CP_Y: 0.717294551185535