Given data and a constraint graph, generate a changepoint output A wrapper around gfpop::gfpop()

Given data and a constraint graph, generate a changepoint output A wrapper around gfpop::gfpop()

generate_changepoint(data_input, graph_input, type = "mean")

Arguments

data_input
vector of data to segment

graph_input
dataframe of class "graph" to constraint the changepoint inference

type
a string defining the cost model to use: "mean", "variance", "poisson", "exp", "negbin"

Value

A dataframe with columns:

changepointa vector of X locations where the changepoints begin

changepoint_enda vector of X locations where the changepoints end

yA vector of Y locations where the changepoints are centered

Examples

data <- gfpop::dataGenerator(10, c(0.1, 0.3, 0.5, 0.8, 1),
c(1, 2, 1, 3, 1), sigma = 1)
graph <- gfpop::graph(type = "std")
generate_changepoint(data, graph)

#>    changepoint changepoint_end           y
#> 1            1               0  0.44630062
#> 2            2               1  2.62898204
#> 3            3               2  4.06502490
#> 4            4               3 -0.63098940
#> 5            5               4  1.51242695
#> 6            6               5  1.13698851
#> 7            7               6  2.47798749
#> 8            8               7  2.94739809
#> 9            9               8  1.54299634
#> 10          10               9  0.08592517