Selects the columns from this graph dataframe that are relevant to gfpop

Selects the columns from this graph dataframe that are relevant to gfpop

select_graph_columns(graph_df)

Arguments

graph_df
the graph to process

Value

a dataframe with 9 columns, compatable with gfpop

Examples

graph <- gfpop::graph(type = "std")
graph$test <- c(1,2)
select_graph_columns(graph)

#>   state1 state2 type parameter penalty   K a min max
#> 1    Std    Std null         1       0 Inf 0  NA  NA
#> 2    Std    Std  std         0       0 Inf 0  NA  NA